+++ next-docs/pages/components/checkbox.tsx
@@ -1,3 +1,4 @@
+import Link from 'next/link';
 import Preview from '../../components/codePreview/Preview';
 import ComponentPageDescription from '../../components/ComponentPageDescription';
 import PropsTable from '../../components/PropsTable';
@@ -13,6 +14,15 @@ const PageAccordion = () => {
   return (
     <>
       <ComponentPageDescription title="Checkbox">
+        <p className="font-medium">
+          NB! The component is depricated. Please use our{' '}
+          <Link href="/core/checkbox">
+            <a className="text-piccolo transition-colors duration-200 hover:text-hit visited:text-hit">
+              new Checkbox
+            </a>
+          </Link>{' '}
+          built with Tailwind.
+        </p>
         <p>
           Checkboxes are used as a list in our forms and signup pages so users
           can select options of their choice. Users are free to choose how many