+++ next-docs/pages/components/loader.tsx
@@ -1,3 +1,4 @@
+import Link from 'next/link';
 import Preview from '../../components/codePreview/Preview';
 import ComponentPageDescription from '../../components/ComponentPageDescription';
 import PropsTable from '../../components/PropsTable';
@@ -11,6 +12,15 @@ const PageLoader = () => {
   return (
     <>
       <ComponentPageDescription title="Loader">
+        <p className="font-medium">
+          NB! The component is depricated. Please use our{' '}
+          <Link href="/core/loader">
+            <a className="text-piccolo transition-colors duration-200 hover:text-hit visited:text-hit">
+              new Loader
+            </a>
+          </Link>{' '}
+          built with Tailwind.
+        </p>
         <p>
           Fondly nicknamed “the launcher”, the Loader ensures users that
           progress is happening so they don't give up and leave the rocket page.