+++ next-docs/public/examples/switchTW/Disabled.tsx
@@ -4,15 +4,8 @@ import { Switch } from '@heathmont/moon-core-tw';
 const Example = () => {
   return (
     <>
-      <Switch
-        checked={true}
-        disabled={true}
-        onChange={() => {
-          console.log('da li radi disabled!');
-        }}
-      />
-
-      <Switch checked={false} disabled={true} onChange={() => {}} />
+      <Switch checked disabled />
+      <Switch checked={false} disabled />
     </>
   );
 };