+++ next-docs/public/examples/switchTW/Form.tsx
@@ -5,16 +5,14 @@ import { Switch } from '@heathmont/moon-core-tw';
 const Example = () => {
   const [state, setState] = useState(true);
   return (
-    <>
-      <form action="/notification-settings" method="post">
-        <Switch
-          checked={state}
-          onChange={setState}
-          name="notification"
-          value="on"
-        />
-      </form>
-    </>
+    <form action="/notification-settings" method="post">
+      <Switch
+        checked={state}
+        onChange={setState}
+        name="notification"
+        value="on"
+      />
+    </form>
   );
 };