+++ packages/fonts/CHANGELOG.md
@@ -1,5 +1,17 @@
 # Change Log
 
+## 9.8.1
+
+### Patch Changes
+
+- Fix code view
+
+## 9.8.0
+
+### Minor Changes
+
+- Deploy fixes
+
 ## 9.7.8
 
 ### Patch Changes